kbuild: Disable stack conservation for GCC
There's plenty of room on the stack for a few more
inlined bytes here and there. The measured stack usage at
runtime is still safe without# Please enter the commit
message for your changes. Lines starting this, and
performance is surely improved at a microscopic level,
so# with '#' will be ignored, and an empty message aborts
the commit. remove it.#
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>#